backend/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
mysqlclient==2.2.7
numpy==2.2.4
packaging==25.0
pyfiglet==1.0.2
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
setuptools==78.1.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0